pyhub/DAY-00-fastapi-project-reviews/sample-project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
jinja2==3.1.3
python-dotenv==1.0.0
httpx==0.27.0
pydantic[email]

